Move common from IntanceVarialbe and LogicalRecord to Field class
Converting from json to pojo works now
